Fix compile on Windows, part 2

+++ sql/multi_range_read.cc
@@ -387,7 +387,7 @@ handler::multi_range_read_info_const(uint keyno, RANGE_SEQ_IF *seq,
       */
       Cost_estimate limit_cost;
       calculate_costs(&limit_cost, keyno, n_ranges,
-                      n_ranges - single_point_ranges,
+                      n_ranges - (uint)single_point_ranges,
                       *flags, top_limit, io_blocks,
                       unassigned_single_point_ranges);
       cost->limit_cost= limit_cost.total_cost();
